+++ src/Web/OAuth/RegisterAPI.hs
@@ -22,12 +22,15 @@
 module Web.OAuth.RegisterAPI where
 
 import Control.Concurrent.MVar
+import Control.Monad (forM_, when)
 import Control.Monad.IO.Class (liftIO)
 import Data.Aeson
 import Data.Map.Strict qualified as Map
 import Data.Maybe (fromMaybe)
 import Data.Text (Text)
+import Data.Text qualified as T
 import GHC.Generics
+import Network.URI qualified as URI
 import Web.OAuth.Types
 import Servant
 
@@ -121,11 +124,12 @@ instance ToJSON ClientRegistrationResponse where
 -- The client_id is generated using a secure random token generator.
 handleRegister :: forall usr. MVar (OAuthState usr) -> ClientRegistrationRequest -> Handler ClientRegistrationResponse
 handleRegister state_var ClientRegistrationRequest{..} = do
+  let default_auth_method = fromMaybe "none" token_endpoint_auth_method
+  validateRedirectUris redirect_uris
   client_id <- ("client_" <>) <$> liftIO generateToken
   let default_grant_types = fromMaybe ["authorization_code", "refresh_token"] grant_types
       default_response_types = fromMaybe ["code"] response_types
       default_scope = fromMaybe "read write" scope
-      default_auth_method = fromMaybe "none" token_endpoint_auth_method
       -- For confidential clients, generate a secret
       secret = if default_auth_method /= "none" then Just <$> generateToken else pure Nothing
   secret' <- liftIO secret
@@ -159,3 +163,45 @@ handleRegister state_var ClientRegistrationRequest{..} = do
       , reg_scope = default_scope
       , reg_token_endpoint_auth_method = default_auth_method
       }
+  where
+    validateRedirectUris :: [Text] -> Handler ()
+    validateRedirectUris uris = do
+      when (null uris) $
+        invalidMetadata "redirect_uris must include at least one absolute URI"
+      forM_ uris $ \uriText -> do
+        parsedUri <- parseAbsolute uriText
+        when (not (null (URI.uriFragment parsedUri))) $
+          invalidMetadata "redirect_uris must not contain URI fragments"
+        case URI.uriScheme parsedUri of
+          "https:" -> ensureAuthority parsedUri
+          "http:" ->
+            case URI.uriAuthority parsedUri of
+              Just auth
+                | isLoopbackHost (URI.uriRegName auth) -> pure ()
+                | otherwise -> invalidMetadata "http redirect_uris are only allowed for loopback clients"
+              Nothing -> invalidMetadata "redirect_uris must include a network host component"
+          _ -> invalidMetadata "redirect_uris must use https scheme"
+
+    parseAbsolute :: Text -> Handler URI.URI
+    parseAbsolute uriText =
+      case URI.parseURI (T.unpack uriText) of
+        Just parsed
+          | not (null (URI.uriScheme parsed)) ->
+              pure parsed
+        _ -> invalidMetadata "redirect_uri is not an absolute URI"
+
+    ensureAuthority :: URI.URI -> Handler ()
+    ensureAuthority parsed =
+      case URI.uriAuthority parsed of
+        Just _ -> pure ()
+        Nothing -> invalidMetadata "redirect_uri must include authority (host)"
+
+    isLoopbackHost :: String -> Bool
+    isLoopbackHost hostName =
+      hostName == "localhost"
+        || hostName == "127.0.0.1"
+        || hostName == "[::1]"
+
+    invalidMetadata :: Text -> Handler a
+    invalidMetadata msg =
+      throwError $ oauthErrorResponse err400 "invalid_client_metadata" (Just msg)